Rename `no-array-push-push` to `prefer-combinable-action`, add supports for `Element#classList.{add,remove}()` and `importScripts()`
