+++ docs/deleted-and-deprecated-rules.md
@@ -6,6 +6,10 @@
 
 Replaced by [`no-instanceof-builtins`](rules/no-instanceof-builtins.md) which covers more cases.
 
+### no-array-push-push
+
+Replaced by [`prefer-combinable-action`](rules/prefer-combinable-action.md) which covers more cases.
+
 ### no-length-as-slice-end
 
 Replaced by [`no-unnecessary-slice-end`](rules/no-unnecessary-slice-end.md) which covers more cases.
