+++ index.js
@@ -13,6 +13,10 @@ const deprecatedRules = createDeprecatedRules({
 		message: 'Replaced by `unicorn/no-unnecessary-slice-end` which covers more cases.',
 		replacedBy: ['unicorn/no-unnecessary-slice-end'],
 	},
+	'no-array-push-push': {
+		message: 'Replaced by `unicorn/prefer-combinable-action` which covers more cases.',
+		replacedBy: ['unicorn/prefer-combinable-action'],
+	},
 });
 
 const externalRules = {
