+++ test/snapshots/prefer-combinable-action.js.md
@@ -1,6 +1,6 @@
-# Snapshot report for `test/no-array-push-push.js`
+# Snapshot report for `test/prefer-combinable-action.js`
 
-The actual snapshot is saved in `no-array-push-push.js.snap`.
+The actual snapshot is saved in `prefer-combinable-action.js.snap`.
 
 Generated by [AVA](https://avajs.dev).
